Move icons to design system
