+++ core/ui/compose/designsystem/src/main/kotlin/app/k9mail/core/ui/compose/designsystem/atom/icon/IconsWithBottomRightOverlay.kt
@@ -1,4 +1,4 @@
-package app.k9mail.core.ui.compose.theme
+package app.k9mail.core.ui.compose.designsystem.atom.icon
 
 import androidx.compose.material.icons.filled.Notifications
 import androidx.compose.material.icons.filled.Person
